new strategy for finding python header files
The strategy we were using produced the wrong location for someone: #9.
There are several possibilities here, and I haven't found a clear
recommendation in the Python docs for what to do. I'm opting to use the
approach that distutils uses to build C extensions, since that is the
implementation that most extensions likely use. See the distutils code
here:

https://github.com/python/cpython/blob/c30098c8c6014f3340a369a31df9c74bdbacc269/Lib/distutils/command/build_ext.py#L150-L151